+++ pyxel/editor/field_cursor.py
@@ -119,6 +119,7 @@ class FieldCursor:
             pyxel.btn(pyxel.KEY_SHIFT)
             or pyxel.btn(pyxel.KEY_CONTROL)
             or pyxel.btn(pyxel.KEY_ALT)
+            or pyxel.btn(pyxel.KEY_SUPER)
         ):
             return